modify configure new type (#84)
* modify configure based on 2.4.0

* add deepCopy func

* add getShardConfig func

* modify the NewShardConfig, use ChainConfig = chainCfg, use reference instead of copy.

* modify grpc test to depend on 2.4.0 version of cfg

* rm unuseful json tag

* modify address from string to account.Address

* modify config module

* modify root height type and networkid type

* modify rootheight type in ShardGenesis struct

* modify rootheight and about this operation

* modify default value of CoinbaseAmount in shardconfig

* modify the type of new

* modify port from uint64 to uint16

+++ cluster/config/slave_config.go
@@ -8,7 +8,7 @@ import (
 
 type SlaveConfig struct {
 	IP            string             `json:"HOST"` // DEFAULT_HOST
-	Port          uint64             `json:"PORT"` // 38392
+	Port          uint16             `json:"PORT"` // 38392
 	ID            string             `json:"ID"`
 	ChainMaskList []*types.ChainMask `json:"-"`
 }
@@ -46,7 +46,7 @@ func (s *SlaveConfig) UnmarshalJSON(input []byte) error {
 func NewDefaultSlaveConfig() *SlaveConfig {
 	slaveConfig := SlaveConfig{
 		IP:   "127.0.0.1",
-		Port: uint64(SlavePort),
+		Port: SlavePort,
 	}
 	return &slaveConfig
 }
